Task1 final

+++ Task1/Program.cs
@@ -18,10 +18,9 @@ namespace Task1
             {
                 Console.WriteLine("Папка существует");
                 Console.WriteLine("Папки:");
-                DateTime OpenDateDir = Directory.GetLastAccessTime(dirName); //Получим время открытия папок
                 string[] dirs = Directory.GetDirectories(dirName);  // Получим все директории в папке
                 foreach (string d in dirs) // Выведем их все
-                    Console.WriteLine("{0} последнее использование {1}", d, OpenDateDir);
+                    Console.WriteLine("{0}", d);
 
                 Console.WriteLine("Файлы:");
 
@@ -31,30 +30,43 @@ namespace Task1
 
                 foreach (FileInfo file in DelFailDir.GetFiles())
                 {
-                    try
-                    {
-                        var timeSpan = file.LastAccessTime;
-                        Console.WriteLine(timeSpan.ToString());
-                        timeSpan.Add(TimeSpan.FromMinutes(30));
-                        file.Delete();
-                    }
-                    catch
-                    {
-                        Console.WriteLine("Нет подходящих для удаления файлов ");
-                    }
-                }
-                foreach (DirectoryInfo dir in DelFailDir.GetDirectories())
-                {
-                    var timeSpan = dir.LastAccessTime;
-                    Console.WriteLine(timeSpan.ToString());
+                    var timeSpan = file.LastAccessTime;
                     var timeSpan1 = DateTime.Now;
                     var timeSpan2 = TimeSpan.FromMinutes(30);
                     var duras = timeSpan1 - timeSpan;
-                    dir.Delete(true);
-
+                    if (duras >= timeSpan2)
+                    { file.Delete(); }
+                    else
+                    {
+                        Console.WriteLine();
+                        Console.WriteLine("Нет файлов,которые не используются более 30 мин.");
+                        Console.WriteLine($"Прошло {duras} времени");
+                        break;
+                    }
                 }
+
+                foreach (DirectoryInfo dir in DelFailDir.GetDirectories())
+                {
+                    var timeSpan = dir.LastAccessTime;
+                    var timeSpan1 = DateTime.Now;
+                    var timeSpan2 = TimeSpan.FromMinutes(30);
+                    var duras = timeSpan1 - timeSpan;
+                    if (duras >= timeSpan2)
+                    {
+                        dir.Delete(true);
+                    }
+                    else
+                    {
+                        Console.WriteLine();
+                        Console.WriteLine("Нет папок,которыене используются более 30 мин");
+                        Console.WriteLine($"Прошло {duras} времени");
+                        break;
+                    }
+                }
+
             }
         }
     }
 }
+
